moving lens components in their own module
