+++ crates/client/src/pages/library.rs
@@ -1,20 +1,14 @@
 use wasm_bindgen::prelude::*;
 use wasm_bindgen_futures::spawn_local;
-use yew::function_component;
 use yew::prelude::*;
 
-use crate::components::{btn::Btn, icons, Header};
+use crate::components::{btn::Btn, icons, lens::LibraryLens, Header};
 use crate::invoke;
 use crate::listen;
 use crate::utils::RequestState;
 use shared::event::ClientEvent;
 use shared::event::ClientInvoke;
-use shared::response::{LensResult, InstallStatus};
-
-#[derive(Properties, PartialEq, Eq)]
-pub struct LensProps {
-    pub result: LensResult,
-}
+use shared::response::LensResult;
 
 async fn fetch_user_installed_lenses() -> Option<Vec<LensResult>> {
     match invoke(ClientInvoke::ListInstalledLenses.as_ref(), JsValue::NULL).await {
@@ -29,60 +23,6 @@ async fn fetch_user_installed_lenses() -> Option<Vec<LensResult>> {
             log::error!("Error fetching lenses: {:?}", e);
             None
         }
-    }
-}
-
-#[function_component(Lens)]
-pub fn lens_component(props: &LensProps) -> Html {
-    let component_styles = classes!(
-        "rounded-md",
-        "bg-neutral-700",
-        "p-4",
-        "text-white",
-        "shadow-md",
-        "overflow-hidden"
-    );
-    let result = &props.result;
-
-    let detail_bar = match &result.progress {
-        InstallStatus::Finished => {
-            html! {
-                <div class="mt-2 text-sm flex flex-row gap-2 items-center">
-                    <a href="https://example.com" class="border-neutral-600 border cursor-pointer font-semibold px-2 py-1 rounded-md text-xs inline-block hover:bg-neutral-600">
-                        {"Details"}
-                    </a>
-                    <a href="https://example.com" class="bg-red-700 cursor-pointer font-semibold px-2 py-1 rounded-md text-xs inline-block hover:bg-red-900">
-                        {"Uninstall"}
-                    </a>
-                </div>
-            }
-        },
-        InstallStatus::Installing { percent, status } => {
-            html! {
-                <div class="mt-2 text-sm">
-                    <div class="text-xs pb-1">{status.clone()}</div>
-                    <div class="w-full bg-stone-800 h-1 rounded-3xl text-xs">
-                        <div class="bg-cyan-400 h-1 rounded-lg pl-2 flex items-center animate-pulse" style={format!("width: {percent}%")}></div>
-                    </div>
-              </div>
-            }
-        }
-    };
-
-    html! {
-        <div class={component_styles}>
-            <div class="mb-1">
-                <div class="text-lg font-semibold">{result.title.to_string()}</div>
-                <div class="text-sm text-neutral-400">
-                    {"Crafted By:"}
-                    <a href={format!("https://github.com/{}", result.author)} target="_blank" class="text-cyan-400">
-                        {format!(" @{}", result.author)}
-                    </a>
-                </div>
-            </div>
-            <div class="text-sm text-neutral-400">{result.description.clone()}</div>
-            {detail_bar}
-        </div>
     }
 }
 
@@ -111,7 +51,7 @@ impl LensManagerPage {
         } else {
             self.user_installed
                 .iter()
-                .map(|data| html! {<Lens result={data.clone()} /> })
+                .map(|data| html! {<LibraryLens result={data.clone()} /> })
                 .collect::<Html>()
         }
     }
